requirements.txt
prefect>=2.14.0
pandas>=2.0.0
azure-storage-blob>=12.19.0
pyarrow>=14.0.0
python-dotenv>=1.0.0

